Convert automation-tools to a library
Add setup.py module and move all tasks to automation_tools.__init__
module. Also import tasks in fabfile to publish the tasks to Fabric.

Now automation_tools could be used as a library or as an utility tool.
